chore(utils): move benchmark and buffer to their respective modules
