+++ tests/utils/test_replay_buffer.py
@@ -21,7 +21,7 @@ import pytest
 import torch
 
 from lerobot.datasets.lerobot_dataset import LeRobotDataset
-from lerobot.utils.buffer import BatchTransition, ReplayBuffer, random_crop_vectorized
+from lerobot.rl.buffer import BatchTransition, ReplayBuffer, random_crop_vectorized
 from tests.fixtures.constants import DUMMY_REPO_ID
